Refactor: Rename package to py_name_entity_recognition (#23)
This commit renames the package from `pyner` to `py_name_entity_recognition` as requested.

The following changes were made:
- Renamed the `pyner` directory to `py_name_entity_recognition`.
- Updated the package name in `pyproject.toml`.
- Updated all import statements and references in the source code, tests, and documentation to use the new package name.
- Updated repository URLs in `pyproject.toml`.
