+++ py_name_entity_recognition/utils/biores_converter.py
@@ -4,8 +4,8 @@ from typing import List, Optional, Tuple
 import spacy
 from spacy.language import Language
 
-from pyner.observability.logging import logger
-from pyner.schemas.core_schemas import BaseEntity
+from py_name_entity_recognition.observability.logging import logger
+from py_name_entity_recognition.schemas.core_schemas import BaseEntity
 
 
 class BIOSESConverter:
