+++ pyproject.toml
@@ -1,12 +1,12 @@
 [tool.poetry]
-name = "pyner"
+name = "py_name_entity_recognition"
 version = "0.1.0"
 description = "LLM-based Named Entity Recognition using LangChain and LangGraph"
 authors = ["Gowtham Rao <rao@ohdsi.org>"]
 license = "Apache-2.0"
 readme = "README.md"
-homepage = "https://github.com/GowthamRao/pyner"
-repository = "https://github.com/GowthamRao/pyner"
+homepage = "https://github.com/GowthamRao/py_name_entity_recognition"
+repository = "https://github.com/GowthamRao/py_name_entity_recognition"
 keywords = ["ner", "named entity recognition", "llm", "langchain", "langgraph", "nlp"]
 
 [tool.poetry.dependencies]
